# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/customer/profile": {
      "get": {
        "operationId": "ProfileController_get",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "security": [{ "cookie": [] }],
        "tags": ["Profile"],
        "x-irent-roles": ["customer"]
      },
      "put": {
        "operationId": "ProfileController_save",
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "description": "Origin frontend dari CORS_ORIGINS.",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "description": "Token dari login/session.",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProfileDto" }
            }
          }
        },
        "responses": { "200": { "description": "" } },
        "security": [{ "cookie": [] }],
        "tags": ["Profile"],
        "x-irent-roles": ["customer"]
      }
    },
    "/api/admin/accounts": {
      "get": {
        "operationId": "AdminAccountsController_accounts",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 1000000,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "type": "integer"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": { "minLength": 1, "maxLength": 100, "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "cookie": [] }],
        "tags": ["AdminAccounts"],
        "x-irent-roles": ["admin"]
      },
      "post": {
        "operationId": "AdminAccountsController_createAdmin",
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "description": "Origin frontend dari CORS_ORIGINS.",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "description": "Token dari login/session.",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          }
        },
        "security": [{ "cookie": [] }],
        "tags": ["AdminAccounts"],
        "x-irent-roles": ["admin"]
      }
    },
    "/api/admin/customers": {
      "get": {
        "operationId": "AdminAccountsController_customersList",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 1000000,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "type": "integer"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": { "minLength": 1, "maxLength": 100, "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "cookie": [] }],
        "tags": ["AdminAccounts"],
        "x-irent-roles": ["admin"]
      }
    },
    "/api/admin/customers/{id}": {
      "get": {
        "operationId": "AdminAccountsController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "cookie": [] }],
        "tags": ["AdminAccounts"],
        "x-irent-roles": ["admin"]
      }
    },
    "/api/admin/customers/{id}/password": {
      "post": {
        "operationId": "AdminAccountsController_reset",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "Origin",
            "in": "header",
            "description": "Origin frontend dari CORS_ORIGINS.",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "description": "Token dari login/session.",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ResetPasswordDto" }
            }
          }
        },
        "responses": { "204": { "description": "" } },
        "security": [{ "cookie": [] }],
        "tags": ["AdminAccounts"],
        "x-irent-roles": ["admin"]
      }
    },
...
  },
  "info": {
    "title": "iRent Semarang API",
    "description": "Rupiah: string desimal; waktu: ISO WIB +07:00. Mutasi: Origin dan X-CSRF-Token; aksi bisnis: Idempotency-Key sesuai backend/API.md. Response rinci mengikuti dokumentasi modul.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": { "type": "apiKey", "in": "cookie", "name": "irent_session" }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 100 },
          "email": { "type": "string", "minLength": 1, "maxLength": 254 },
          "phone": { "type": "string", "minLength": 1, "maxLength": 40 },
          "password": { "type": "string", "minLength": 12, "maxLength": 128 }
        },
        "required": ["name", "password"]
      },
...
      "ProfileDto": {
        "type": "object",
        "properties": {
          "fullName": { "type": "string", "minLength": 1, "maxLength": 150 },
          "address": { "type": "string", "minLength": 1, "maxLength": 1000 },
          "phoneActive": { "type": "string", "minLength": 1, "maxLength": 40 },
          "nik": { "type": "string", "pattern": "^[0-9]{16}$" },
          "phoneAlt": { "type": "string", "nullable": true, "maxLength": 40 },
          "instagram": { "type": "string", "nullable": true, "maxLength": 100 },
          "emailContact": {
            "type": "string",
            "nullable": true,
            "maxLength": 254
          }
        },
        "required": ["fullName", "address", "phoneActive"]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "password": { "type": "string", "minLength": 12, "maxLength": 128 }
        },
        "required": ["password"]
      },
...
    }
  }
}